requirements.txt
Django
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
django-environ
drf-spectacular
django-filter==23.2.0
pytest
pytest-django

